# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Farmer Market Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /markets:
    get:
      operationId: listMarkets
      tags: [markets]
      summary: List all markets
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Market"
  /markets/recommend:
    get:
      operationId: recommendMarkets
      tags: [markets]
      summary: Recommend best markets for a crop
      parameters:
        - name: crop
          in: query
          required: true
          schema:
            type: string
        - name: location
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recommended markets ranked by price
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketRecommendation"
...
  /markets/nearby:
    get:
      operationId: getNearbyMarkets
      tags: [markets]
      summary: Get markets near a GPS coordinate
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Nearby markets sorted by distance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NearbyMarket"
...
components:
  schemas:
...
    Market:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
        district:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        type:
          type: string
      required:
        - id
        - name
        - state
        - district
        - latitude
        - longitude
        - type
    MarketRecommendation:
      type: object
      properties:
        market:
          type: string
        state:
          type: string
        district:
          type: string
        modalPrice:
          type: number
        mspPrice:
          type: ["number", "null"]
        premiumOverMsp:
          type: ["number", "null"]
        distanceKm:
          type: ["number", "null"]
        score:
          type: number
        reason:
          type: string
      required:
        - market
        - state
        - district
        - modalPrice
        - mspPrice
        - premiumOverMsp
        - distanceKm
        - score
        - reason
...
    NearbyMarket:
      type: object
      required: [id, name, state, district, type, distanceKm]
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
        district:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        type:
          type: string
        distanceKm:
          type: number
...